+++ PrimeFactorizer/PrimeFactor.h
@@ -11,13 +11,13 @@ public:
 	PrimeFactor() ;//no lib
 	PrimeFactor(shared_ptr <PrimeLibrary> l);//0-4; select library size with 0 being no library and 4 being the largest
 	~PrimeFactor();
-	vector <double> factor(double num);
+	vector <uint64_t> factor(uint64_t num);
 
 private:
 
 	unique_ptr <PrimeCheck> Check;
 	unique_ptr <PrimeDivisor> Div;
-	vector <double> Divisors;
+	vector <uint64_t> Divisors;
 
 };
